+++ test/mixed.js
@@ -400,3 +400,21 @@ describe('uuid1', function () {
         assert.equal(result, false);
     });
 });
+
+describe('uuid', function () {
+    it('#1 should be return true', function () {
+        var result = be.uuid('9e3a0460-d72d-11e4-a631-c8e0eb141dab');
+        console.log(result);
+        assert.equal(result, true);
+    });
+    it('#2 should be return false', function () {
+        var result = be.uuid('366a77ba-d506-4a03-a730-318b8e6be8c5');
+        console.log(result);
+        assert.equal(result, true);
+    });
+    it('#3 should be return true', function () {
+        var result = be.uuid('39888f87-fb62-5988-a425-b2ea63f5b81e');
+        console.log(result);
+        assert.equal(result, true);
+    });
+});