+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.7.0] - 2017-07-22
 ### Added
+- `be.uuid`, detect if is UUID
 - `be.uuid1`, detect if is UUID v1
 - `be.uuid3`, detect if is UUID v3
 - `be.uuid4`, detect if is UUID v4