+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `be.uuid3`, detect if is UUID v3
 - `be.uuid4`, detect if is UUID v4
 - `be.uuid5`, detect if is UUID v5
+- `be.isrc`, detect if is ISRC
 
 ## [1.6.3] - 2017-07-20
 ### Added