+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.7.1] - 2017-07-22
+### Fixed
+- Wrong documentation
+
 ## [1.7.0] - 2017-07-22
 ### Added
 - `be.uuid`, detect if is UUID